update hamburger menu and heading
