+++ frontend/src/components/FormContainer.jsx
@@ -1,11 +1,11 @@
 const FormContainer = ({ children }) => {
   return (
     <table>
-        <tr className='justify-content-md-center mt-5'>
-            <td className='card p-5'>
-                {children}
-            </td>
-        </tr>
+      <tr className='justify-content-md-center mt-2'>
+        <td className='card p-5'>
+          {children}
+        </td>
+      </tr>
     </table>
   )
 }
